Dropped boost and switched to C-style pthread/mutex shared memory
So the code in welcome.cc still has my boost code, I've left it alone
and started semi-fresh in welcome_c.cc. In that new file, I've
successfully gotten C-style (shm_open(), mmap(), etc) to work. I added
pthread mutexes to both read and write, but not sure if that's actually
properly locking the shared memory.

I wrote sample code that writes "Hello world!" and an integer counting 
up incrementally every second to the shared memory, and a reader that 
reads the memory content and prints it every second as well.

Maybe I'll stick with this? If I have time, I'll look into Qt's shared
memory functionalities and see if that can work the way I'd like it to.