Add functional maya prototype plug-in
Inside of the /stringless-maya directory is a functioning Maya 2016.5
SP2 plug-in (and successfully building Netbeans project) built using the
Maya 2016.5 SP2 Qt library. To use the plug-in, load the compiled 
libstringless.so plug-in file into Maya and execute the "stringless" MEL
command to see the demo window.

The Netbeans pre-build/build configurations were heavily modified to
support Maya's custom version of Qt's qmake makefile creator. In a 
nutshell, Maya qmake uses the stringless.pro file to find the source
files and library files for Maya and Maya's Qt, and generates the proper
Makefile inside build/Debug. 

Most of the config files should work for any platform, but the qmake for
a given Maya version should be used and the MAYA_LOCATION variable in
the userconfig file should be updated to the correct path.