Possible fix for issues with ui file loading
I've been trying to load the stringless.ui file in my plug-in
with lots of trouble relating to relative file paths; specifically,
I couldn't find the path to the plug-in itself at runtime (paths
were only coming up respective to Maya, understandably).

This change in the Qt project file may be the fix, I have to try
compiling with the new local version before I'll push the working
version. Will revert this if it doesn't.