Begin initial communication between server and Maya plug-in
The Maya plug-in is now reading from the shared memory and outputting 
the x value of point 0 to a label, upon each click of the "Start"
button. This is done by modifying the stringless standalone compile
process to build certain classes as shared libraries, then linking the
Maya plug-in to those libraries.

The point data needs to be formatted into a Qt Model Table and then sent
to the Table View.

Project scope notes: I'm going to keep with just one frame capture at
a time, and worry about the realtime portion later.