Maya plug-in pre-alpha
There is a lot to say about this commit; it's a sloppy, ugly mess, for 
one. But very satisfyingly, the Maya plug-in works, at least for one
point for now.

The plug-in has been re-designed from the ground up, I'm no longer using
Qt (no need, not developing any custom UI), and am now using a mixture
of a Maya Threaded Device Node, a Maya Cmd, and a MEL script to power
the plug-in. The ThreadedDeviceNode handles the shared memory accessing
and data transfer, the Cmd registers the MEL command to initialize the
plug-in, and the MEL script instantiates the device, custom rigging, and
connects the nodes of the rig to the device data.

There's a lot to do in terms of redesigning the incoming data structure,
adding support for all the other 68 data points, implementing depth
along the Z axis, code cleanup across all plug-in files, and especially
cleaning up the mess of my compile commands (don't look at those
NetBeans project files too hard).

My top priority is getting all the points running, after which I'll begin
to refactor the plug-in project to be more open-source friendly (and
merge to master).

But this is a major milestone for myself, so hooray! :)