Add landmark sample per frame arg
Allow user to specify more than one landmark sampling per frame, in the
attempts to get higher quality captures. Honestly not sure if it does
yet, but I hoped it would help.

Also been fooling around with blurring the image in the FaceDetector,
since my initial theory of all the noise in the capture data being a 
product of the noisy input image. Will need to test with a higher fps
camera to see if it really helps.